Create testPoints to test all Point related classes and their computations. Moved Station test from testFrames here as the first case.
